# medium_api_gateway/api/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.Comment:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.CommentUser'
      user_id:
        type: integer
    type: object
  models.CommentUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      profile_image_url:
        type: string
    type: object
...
  models.CreateCommentRequest:
    properties:
      description:
        type: string
      post_id:
        type: integer
    type: object
  models.CreateOrUpdateLikeRequest:
    properties:
      post_id:
        type: integer
      status:
        type: boolean
    required:
    - post_id
    type: object
...
  models.GetAllCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      count:
        type: integer
    type: object
...
  models.Like:
    properties:
      id:
        type: integer
      post_id:
        type: integer
      status:
        type: boolean
      user_id:
        type: integer
    type: object
  models.LikesAndDislikesCount:
    properties:
      dislikes:
        type: integer
      likes:
        type: integer
    type: object
...
  models.ResponseError:
    properties:
      error:
        type: string
    type: object
  models.ResponseSuccess:
    properties:
      success:
        type: string
    type: object
  models.UpdateComment:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.UpdateCommentRequest:
    properties:
      description:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a blog service api.
  version: "2.0"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Get comments
      parameters:
      - default: 10
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        in: query
        name: page
        required: true
        type: integer
      - in: query
        name: post_id
        type: integer
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GetAllCommentsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get comments
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create a comment
      parameters:
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a comment
      tags:
      - comment
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: Update comment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UpdateComment'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update comment
      tags:
      - comment
  /likes:
    post:
      consumes:
      - application/json
      description: Create Or Update like
      parameters:
      - description: like
        in: body
        name: like
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrUpdateLikeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Like'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create Or Update like
      tags:
      - like
  /likes/user-post-likes/{id}:
    get:
      consumes:
      - application/json
      description: Get likes and dislikes count by giving to query post_id
      parameters:
      - description: post_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LikesAndDislikesCount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get likes and dislike count by giving to query post_id
      tags:
      - like
  /likes/user-post/{id}:
    get:
      consumes:
      - application/json
      description: Get like by giving to query post_id
      parameters:
      - description: post_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Like'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get like by giving to query post_id
      tags:
      - like
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"